Importer, split all animation into blender own files
