+++ addons/io_scene_gltf2/blender/imp/gltf2_blender_animation_bone.py
@@ -18,26 +18,26 @@
  * Contributor(s): Julien Duroure.
  *
  * ***** END GPL LICENSE BLOCK *****
- * This development is done in strong collaboration with Airbus Defence & Space
  """
 
 import bpy
 from mathutils import Quaternion, Matrix
 
-class AnimationBone():
-    def __init__(self, animation):
-        self.animation = animation
+from .gltf2_blender_animation_data import *
 
-    def anim(self):
-        obj   = bpy.data.objects[self.animation.gltf.skins[self.animation.node.skin_id].blender_armature_name]
-        bone  = obj.pose.bones[self.animation.node.blender_bone_name]
+class BlenderBoneAnim():
+
+    @staticmethod
+    def anim(pyanim):
+        obj   = bpy.data.objects[pyanim.animation.gltf.skins[pyanim.animation.node.skin_id].blender_armature_name]
+        bone  = obj.pose.bones[pyanim.animation.node.blender_bone_name]
         fps = bpy.context.scene.render.fps
 
-        for anim in self.animation.anims.keys():
-            if self.animation.gltf.animations[anim].name:
-                name = self.animation.gltf.animations[anim].name + "_" + obj.name
+        for anim in pyanim.animation.anims.keys():
+            if pyanim.animation.gltf.animations[anim].name:
+                name = pyanim.animation.gltf.animations[anim].name + "_" + obj.name
             else:
-                name = "Animation_" + str(self.animation.gltf.animations[anim].index) + "_" + obj.name
+                name = "Animation_" + str(pyanim.animation.gltf.animations[anim].index) + "_" + obj.name
             if name not in bpy.data.actions:
                 action = bpy.data.actions.new(name)
             else:
@@ -46,63 +46,63 @@ class AnimationBone():
                 obj.animation_data_create()
             obj.animation_data.action = bpy.data.actions[action.name]
 
-            for channel in self.animation.anims[anim]:
+            for channel in pyanim.animation.anims[anim]:
                 if channel.path == "translation":
                     blender_path = "location"
                     for key in channel.data:
-                        transform = Matrix.Translation(self.animation.gltf.convert.location(list(key[1])))
-                        if not self.animation.node.parent:
+                        transform = Matrix.Translation(pyanim.animation.gltf.convert.location(list(key[1])))
+                        if not pyanim.animation.node.parent:
                             mat = transform
                         else:
-                            if not self.animation.gltf.scene.nodes[self.animation.node.parent].is_joint: # TODO if Node in another scene
-                                parent_mat = bpy.data.objects[self.animation.gltf.scene.nodes[self.animation.node.parent].blender_object].matrix_world
+                            if not pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].is_joint: # TODO if Node in another scene
+                                parent_mat = bpy.data.objects[pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].blender_object].matrix_world
                                 mat = transform
                             else:
-                                parent_mat = self.animation.gltf.scene.nodes[self.animation.node.parent].blender_bone_matrix
+                                parent_mat = pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].blender_bone_matrix
 
                                 mat = (parent_mat.to_quaternion() * transform.to_quaternion()).to_matrix().to_4x4()
                                 mat = Matrix.Translation(parent_mat.to_translation() + ( parent_mat.to_quaternion() * transform.to_translation() )) * mat
                                 #TODO scaling of bones ?
 
-                        bone.location = self.animation.node.blender_bone_matrix.inverted() * mat.to_translation()
+                        bone.location = pyanim.animation.node.blender_bone_matrix.inverted() * mat.to_translation()
                         bone.keyframe_insert(blender_path, frame = key[0] * fps, group='location')
 
 
                     # Setting interpolation
                     for fcurve in [curve for curve in obj.animation_data.action.fcurves if curve.group.name == "rotation"]:
                         for kf in fcurve.keyframe_points:
-                            self.animation.set_interpolation(channel.interpolation, kf)
+                            BlenderAnimationData.set_interpolation(channel.interpolation, kf)
 
                 elif channel.path == "rotation":
                     blender_path = "rotation_quaternion"
                     for key in channel.data:
-                        transform = (self.animation.gltf.convert.quaternion(key[1])).to_matrix().to_4x4()
-                        if not self.animation.node.parent:
+                        transform = (pyanim.animation.gltf.convert.quaternion(key[1])).to_matrix().to_4x4()
+                        if not pyanim.animation.node.parent:
                             mat = transform
                         else:
-                            if not self.animation.gltf.scene.nodes[self.animation.node.parent].is_joint: # TODO if Node in another scene
-                                parent_mat = bpy.data.objects[self.animation.gltf.scene.nodes[self.animation.node.parent].blender_object].matrix_world
+                            if not pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].is_joint: # TODO if Node in another scene
+                                parent_mat = bpy.data.objects[pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].blender_object].matrix_world
                                 mat = transform
                             else:
-                                parent_mat = self.animation.gltf.scene.nodes[self.animation.node.parent].blender_bone_matrix
+                                parent_mat = pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].blender_bone_matrix
 
                                 mat = (parent_mat.to_quaternion() * transform.to_quaternion()).to_matrix().to_4x4()
                                 mat = Matrix.Translation(parent_mat.to_translation() + ( parent_mat.to_quaternion() * transform.to_translation() )) * mat
                                 #TODO scaling of bones ?
 
-                        bone.rotation_quaternion = self.animation.node.blender_bone_matrix.to_quaternion().inverted() * mat.to_quaternion()
+                        bone.rotation_quaternion = pyanim.animation.node.blender_bone_matrix.to_quaternion().inverted() * mat.to_quaternion()
                         bone.keyframe_insert(blender_path, frame = key[0] * fps, group='rotation')
 
                     # Setting interpolation
                     for fcurve in [curve for curve in obj.animation_data.action.fcurves if curve.group.name == "rotation"]:
                         for kf in fcurve.keyframe_points:
-                            self.animation.set_interpolation(channel.interpolation, kf)
+                            BlenderAnimationData.set_interpolation(channel.interpolation, kf)
 
 
                 elif channel.path == "scale":
                     blender_path = "scale"
                     for key in channel.data:
-                        s = self.animation.gltf.convert.scale(list(key[1]))
+                        s = pyanim.animation.gltf.convert.scale(list(key[1]))
                         transform = Matrix([
                             [s[0], 0, 0, 0],
                             [0, s[1], 0, 0],
@@ -110,14 +110,14 @@ class AnimationBone():
                             [0, 0, 0, 1]
                         ])
 
-                        if not self.animation.node.parent:
+                        if not pyanim.animation.node.parent:
                             mat = transform
                         else:
-                            if not self.animation.gltf.scene.nodes[self.animation.node.parent].is_joint: # TODO if Node in another scene
-                                parent_mat = bpy.data.objects[self.animation.gltf.scene.nodes[self.animation.node.parent].blender_object].matrix_world
+                            if not pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].is_joint: # TODO if Node in another scene
+                                parent_mat = bpy.data.objects[pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].blender_object].matrix_world
                                 mat = transform
                             else:
-                                parent_mat = self.animation.gltf.scene.nodes[self.animation.node.parent].blender_bone_matrix
+                                parent_mat = pyanim.animation.gltf.scene.nodes[pyanim.animation.node.parent].blender_bone_matrix
                                 mat = parent_mat.inverted() * transform
 
 
@@ -128,4 +128,4 @@ class AnimationBone():
                     # Setting interpolation
                     for fcurve in [curve for curve in obj.animation_data.action.fcurves if curve.group.name == "rotation"]:
                         for kf in fcurve.keyframe_points:
-                            self.animation.set_interpolation(channel.interpolation, kf)
+                            BlenderAnimationData.set_interpolation(channel.interpolation, kf)
