+++ addons/io_scene_gltf2/blender/imp/gltf2_blender_animation_node.py
@@ -18,67 +18,66 @@
  * Contributor(s): Julien Duroure.
  *
  * ***** END GPL LICENSE BLOCK *****
- * This development is done in strong collaboration with Airbus Defence & Space
  """
 
 import bpy
+from mathutils import Quaternion, Matrix, Vector
 
-from mathutils import Vector, Quaternion
+from .gltf2_blender_animation_data import *
 
-class AnimationNode():
-    def __init__(self, animation):
-        self.animation = animation
+class BlenderNodeAnim():
 
-    def anim(self):
-        obj = bpy.data.objects[self.animation.node.blender_object]
+    @staticmethod
+    def anim(pyanim):
+        obj = bpy.data.objects[pyanim.animation.node.blender_object]
         fps = bpy.context.scene.render.fps
 
-        for anim in self.animation.anims.keys():
-            if self.animation.gltf.animations[anim].name:
-                name = self.animation.gltf.animations[anim].name + "_" + obj.name
+        for anim in pyanim.animation.anims.keys():
+            if pyanim.animation.gltf.animations[anim].name:
+                name = pyanim.animation.gltf.animations[anim].name + "_" + obj.name
             else:
-                name = "self.animation_" + str(self.animation.gltf.animations[anim].index) + "_" + obj.name
+                name = "pyanim.animation_" + str(pyanim.animation.gltf.animations[anim].index) + "_" + obj.name
             action = bpy.data.actions.new(name)
             if not obj.animation_data:
                 obj.animation_data_create()
             obj.animation_data.action = bpy.data.actions[action.name]
 
-            for channel in self.animation.anims[anim]:
+            for channel in pyanim.animation.anims[anim]:
                 if channel.path in ['translation', 'rotation', 'scale']:
 
                     if channel.path == "translation":
                         blender_path = "location"
                         for key in channel.data:
-                           obj.location = Vector(self.animation.gltf.convert.location(list(key[1])))
+                           obj.location = Vector(pyanim.animation.gltf.convert.location(list(key[1])))
                            obj.keyframe_insert(blender_path, frame = key[0] * fps, group='location')
 
                         # Setting interpolation
                         for fcurve in [curve for curve in obj.animation_data.action.fcurves if curve.group.name == "rotation"]:
                             for kf in fcurve.keyframe_points:
-                                self.animation.set_interpolation(channel.interpolation, kf)
+                                BlenderAnimationData.set_interpolation(channel.interpolation, kf)
 
                     elif channel.path == "rotation":
                         blender_path = "rotation_quaternion"
                         for key in channel.data:
-                            obj.rotation_quaternion = self.animation.gltf.convert.quaternion(key[1])
+                            obj.rotation_quaternion = pyanim.animation.gltf.convert.quaternion(key[1])
                             obj.keyframe_insert(blender_path, frame = key[0] * fps, group='rotation')
 
                         # Setting interpolation
                         for fcurve in [curve for curve in obj.animation_data.action.fcurves if curve.group.name == "rotation"]:
                             for kf in fcurve.keyframe_points:
-                                self.animation.set_interpolation(channel.interpolation, kf)
+                                BlenderAnimationData.set_interpolation(channel.interpolation, kf)
 
 
                     elif channel.path == "scale":
                         blender_path = "scale"
                         for key in channel.data:
-                            obj.scale = Vector(self.animation.gltf.convert.scale(list(key[1])))
+                            obj.scale = Vector(pyanim.animation.gltf.convert.scale(list(key[1])))
                             obj.keyframe_insert(blender_path, frame = key[0] * fps, group='scale')
 
                         # Setting interpolation
                         for fcurve in [curve for curve in obj.animation_data.action.fcurves if curve.group.name == "rotation"]:
                             for kf in fcurve.keyframe_points:
-                                self.animation.set_interpolation(channel.interpolation, kf)
+                                BlenderAnimationData.set_interpolation(channel.interpolation, kf)
 
                 elif channel.path == 'weights':
                     cpt_sk = 0
