+++ addons/io_scene_gltf2/io/com/gltf2_io_animation_data.py
@@ -21,41 +21,19 @@
  * This development is done in strong collaboration with Airbus Defence & Space
  """
 
-from .animation_node import *
-from .animation_bone import *
+from .gltf2_io_animation_node import *
+from .gltf2_io_animation_bone import *
 
 class AnimationData():
     def __init__(self, node, gltf):
         self.node = node
         self.gltf = gltf
         self.anims = {}
-        self.node_anim = AnimationNode(self)
-        self.bone_anim = AnimationBone(self)
+        self.node_anim = PyAnimationNode(self)
+        self.bone_anim = PyAnimationBone(self)
 
 
     def set_anim(self, channel):
         if channel.anim.index not in self.anims.keys():
             self.anims[channel.anim.index] = []
         self.anims[channel.anim.index].append(channel)
-
-    def blender_anim(self):
-        if self.node.is_joint:
-            self.bone_anim.anim()
-        else:
-            self.node_anim.anim()
-
-        for child in self.node.children:
-            child.animation.blender_anim()
-
-    def set_interpolation(self, interpolation, kf):
-        if interpolation == "LINEAR":
-            kf.interpolation = 'LINEAR'
-        elif interpolation == "STEP":
-            kf.interpolation = 'CONSTANT'
-        elif interpolation == "CATMULLROMSPLINE":
-            kf.interpolation = 'BEZIER' #TODO
-        elif interpolation == "CUBICSPLINE":
-            kf.interpolation = 'BEZIER' #TODO
-        else:
-            self.gltf.log.error("Unknown interpolation : " + self.interpolation)
-            kf.interpolation = 'BEZIER'
