+++ addons/io_scene_gltf2/io/com/gltf2_io_animation_node.py
@@ -1,0 +1,26 @@
+"""
+ * ***** BEGIN GPL LICENSE BLOCK *****
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ *
+ * Contributor(s): Julien Duroure.
+ *
+ * ***** END GPL LICENSE BLOCK *****
+ * This development is done in strong collaboration with Airbus Defence & Space
+ """
+
+class PyAnimationNode():
+    def __init__(self, animation):
+        self.animation = animation
