+++ addons/io_scene_gltf2/io/com/gltf2_io_animation_sampler.py
@@ -21,7 +21,7 @@
  * This development is done in strong collaboration with Airbus Defence & Space
  """
 
-from ....io.com.gltf2_io_accessor import *
+from .gltf2_io_accessor import *
 
 class Sampler():
     def __init__(self, index, json, gltf, channels=0):
